messy state
realized that we can't implement backward_pass1 in terms of backward_pass2
because bp2 wants to know which refs are lin vs nonlin up-front, whereas bp1
callers don't tell it that and it just goes with the flow

but we like bp2 because:
* caller contract is clearer
* GradAccum gives us useful laziness
* GradAccum bp2 is simpler: no dangling refs, more unified
* less confusing for vjp3 (we're not sure if it's _needed_ for vjp3)

so the plan is to , instead of implementing bp1 in terms of bp2, just switch
everything to bp2! then we can delete bp1. fixit week!